Mayor refactor on notifications and solving bugs
