+++ mobileapp/redux/store.js
@@ -1,8 +1,10 @@
 import { configureStore } from "@reduxjs/toolkit";
 import userSlice from "./slices/userSlice";
+import executingSlice from "./slices/executingSlice";
 
 export default configureStore({
     reducer: {
-        user: userSlice
+        user: userSlice,
+        executing: executingSlice
     }
 });
